completed userDTO integration in controler and service

+++ src/main/java/com/LavaCandy/Personal/Task/Manager/service/UserService.java
@@ -2,6 +2,9 @@ package com.LavaCandy.Personal.Task.Manager.service;
 
 import com.LavaCandy.Personal.Task.Manager.model.User;
 import com.LavaCandy.Personal.Task.Manager.repository.UserRepository;
+import com.LavaCandy.Personal.Task.Manager.dto.UserRequestDTO;
+import com.LavaCandy.Personal.Task.Manager.dto.UserResponseDTO;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -14,30 +17,50 @@ public class UserService {
     @Autowired
     private UserRepository userRepository;
     
-    public User creatUser(User user) {
-        return userRepository.save(user);
+    // public User creatUser(User user) {
+    //     return userRepository.save(user);
+    // }
+
+    public UserResponseDTO createUser(UserRequestDTO userRequestDTO) {
+        User user = new User();
+        user.setName(userRequestDTO.getName());
+        user.setEmail(userRequestDTO.getEmail());
+        user.setPassword(userRequestDTO.getPassword());
+        User savedUser = userRepository.save(user);
+              
+        return new UserResponseDTO(savedUser);
     }
 
-    public List<User> getAllUser() {
-        return userRepository.findAll();
+    public List<UserResponseDTO> getAllUsers() {
+        return userRepository.findAll().stream().map(UserResponseDTO::new).toList();
     }
 
-    public User getUserById(Long id) {
-        return userRepository.findById(id).orElse(null);
+    public UserResponseDTO getUserById(Long id) {
+        User foundUser =  userRepository.findById(id).orElse(null);
+        if (foundUser != null) {
+            return new UserResponseDTO(foundUser);
+        } else return null;
     }
 
-    public void deleteUser(Long id) {
-        userRepository.deleteById(id);
+    public boolean deleteUser(Long id) {
+        if (userRepository.existsById(id)) {
+            userRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 
-    public User updateUser(Long id, User user) {
+
+    public UserResponseDTO updateUser(Long id, UserRequestDTO userRequestDTO) {
         Optional<User> existingUser = userRepository.findById(id);
         if(existingUser.isPresent()) {
             User updatedUser = existingUser.get();
-            updatedUser.setName(user.getName());
-            updatedUser.setEmail(user.getEmail());
-            updatedUser.setPassword(user.getPassword());
-            return userRepository.save(updatedUser);
+            updatedUser.setName(userRequestDTO.getName());
+            updatedUser.setEmail(userRequestDTO.getEmail());
+            updatedUser.setPassword(userRequestDTO.getPassword());
+            // Save the updated user
+            User updatedUser2 =  userRepository.save(updatedUser);
+            return new UserResponseDTO(updatedUser2);
         }
         return null;
     }
